+++ spines/bullet_spine.cpp
@@ -14,7 +14,6 @@
 
 #include "upkie/cpp/actuation/BulletInterface.h"
 #include "upkie/cpp/model/joints.h"
-#include "upkie/cpp/model/servo_layout.h"
 #include "upkie/cpp/observers/BaseOrientation.h"
 #include "upkie/cpp/observers/FloorContact.h"
 #include "upkie/cpp/observers/ObserverPipeline.h"
@@ -229,7 +228,6 @@ int main(const char* argv0, const CommandLineArguments& args) {
   // a "b3AlignedObjectArray reserve out-of-memory" error below.
 
   // Simulator
-  const auto servo_layout = upkie::cpp::model::servo_layout();
   const double base_altitude = args.space ? 0.0 : 0.6;  // [m]
   BulletInterface::Parameters bullet_params(Dictionary{});
   bullet_params.argv0 = argv0;
@@ -246,7 +244,7 @@ int main(const char* argv0, const CommandLineArguments& args) {
         "external/upkie_description/urdf/upkie_" + args.robot_variant + ".urdf";
   }
   bullet_params.env_urdf_paths = args.extra_urdf_paths;
-  BulletInterface interface(servo_layout, bullet_params);
+  BulletInterface interface(bullet_params);
 
   // Spine
   Spine::Parameters spine_params;