+++ upkie/cpp/actuation/BulletInterface.cpp
@@ -33,9 +33,8 @@ std::string find_plane_urdf(const std::string argv0) {
       "upkie/upkie/cpp/actuation/bullet/plane/plane.urdf");
 }
 
-BulletInterface::BulletInterface(const ServoLayout& layout,
-                                 const Parameters& params)
-    : Interface(layout), params_(params), rng_(std::random_device()()) {
+BulletInterface::BulletInterface(const Parameters& params)
+    : Interface(), params_(params), rng_(std::random_device()()) {
   // Start simulator
   auto flag = (params.gui ? eCONNECT_GUI : eCONNECT_DIRECT);
   bool is_connected = bullet_.connect(flag);