+++ upkie/cpp/actuation/MockInterface.h
@@ -21,10 +21,9 @@ class MockInterface : public Interface {
  public:
   /*! Create mock actuator interface.
    *
-   * \param[in] layout Servo layout.
    * \param[in] dt Simultation timestep in [s].
    */
-  MockInterface(const ServoLayout& layout, const double dt);
+  MockInterface(const double dt);
 
   //! Default destructor
   ~MockInterface() = default;